del serve_logs and move sync() to backend
